analytics-python/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Data Science
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3